fix: v2 bugs
