+++ cmd.go
@@ -28,6 +28,9 @@ import (
 	"github.com/spf13/cobra"
 	"go.ketch.com/lib/orlop/v2/errors"
 	"go.ketch.com/lib/orlop/v2/log"
+	"go.ketch.com/lib/orlop/v2/logging"
+	"go.ketch.com/lib/orlop/v2/service"
+	"go.ketch.com/lib/orlop/v2/version"
 	"go.opentelemetry.io/contrib/instrumentation/runtime"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/exporters/prometheus"
@@ -39,6 +42,7 @@ import (
 	selector "go.opentelemetry.io/otel/sdk/metric/selector/simple"
 	"go.opentelemetry.io/otel/sdk/resource"
 	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
+	"go.opentelemetry.io/otel/trace"
 	"go.uber.org/fx"
 	stdlog "log"
 	"os"
@@ -214,6 +218,11 @@ func (r *Runner) runE(runner interface{}, cfg interface{}) func(cmd *cobra.Comma
 
 		l := log.New()
 
+		loglevelFlag, err := cmd.Flags().GetString("loglevel")
+		if err != nil {
+			return err
+		}
+
 		if module, ok := runner.(fx.Option); ok {
 			if _, ok = cfg.(ProvidesFxOptions); ok {
 				runner = func(ctx context.Context, cfg ProvidesFxOptions) error {
@@ -221,6 +230,12 @@ func (r *Runner) runE(runner interface{}, cfg interface{}) func(cmd *cobra.Comma
 						log.FxLogger(l),
 						FxOptions(cfg),
 						FxContext(ctx),
+						fx.Supply(service.Name(r.prefix)),
+						fx.Supply(logging.Level(loglevelFlag)),
+						fx.Provide(func(tracerProvider trace.TracerProvider) trace.Tracer {
+							return tracerProvider.Tracer(version.Name, trace.WithInstrumentationVersion(version.Name))
+						}),
+						Module,
 						module,
 					)
