+++ parameter/module.go
@@ -22,8 +22,4 @@ package parameter
 
 import "go.uber.org/fx"
 
-var Module = fx.Options(
-	fx.Provide(
-		StoreFromObjectStore,
-	),
-)
+var Module = fx.Options()
