+++ telemetry/module.go
@@ -24,7 +24,6 @@ import "go.uber.org/fx"
 
 var Module = fx.Options(
 	fx.Provide(
-		tracer,
-		metrics,
+		telemetry,
 	),
 )
